+++ src/components/cast-on-tv/cast-before-play.js
@@ -6,6 +6,7 @@ import {IconType} from '../icon/index';
 import {actions} from '../../reducers/backdrop';
 import {Icon} from '../icon/icon';
 import {Localizer, Text} from 'preact-i18n';
+import {KeyMap} from '../../utils/key-map';
 import {withPlayer} from '../player';
 import {withLogger} from 'components/logger';
 
@@ -84,17 +85,26 @@ class CastBeforePlay extends Component {
       }
       return (
         <div>
-          <div className={rootStyle.join(' ')} onClick={() => this.onClick()}>
-            <a className={[style.btn, style.btnDarkTransparent, style.castOnTvButton].join(' ')}>
-              <div className={style.castOnTvIconContainer}>
-                <Icon type={props.icon} />
-              </div>
-              <Localizer>
+          <div className={rootStyle.join(' ')}>
+            <Localizer>
+              <button
+                tabIndex="0"
+                aria-label={<Text id={'cast.play_on_tv'} />}
+                onClick={() => this.onClick()}
+                onKeyDown={e => {
+                  if (e.keyCode === KeyMap.ENTER) {
+                    this.onClick();
+                  }
+                }}
+                className={[style.btn, style.btnDarkTransparent, style.castOnTvButton].join(' ')}>
+                <div className={style.castOnTvIconContainer}>
+                  <Icon type={props.icon} />
+                </div>
                 <span>
                   <Text id="cast.play_on_tv" />
                 </span>
-              </Localizer>
-            </a>
+              </button>
+            </Localizer>
           </div>
         </div>
       );