+++ src/components/picture-in-picture-overlay/picture-in-picture-overlay.js
@@ -74,7 +74,7 @@ class PictureInPictureOverlay extends Component {
               </span>
             </Localizer>
             <Localizer>
-              <button className={[style.pictureInPictureButton, style.controlButtonContainer].join(' ')} onClick={() => this._handleClick()}>
+              <button tabIndex="0" className={[style.pictureInPictureButton, style.controlButton].join(' ')} onClick={() => this._handleClick()}>
                 <Text id="pictureInPicture.overlay_button" />
               </button>
             </Localizer>