+++ src/components/shell/shell.js
@@ -214,7 +214,8 @@ class Shell extends Component {
    */
   componentDidMount() {
     const {player, forceTouchUI} = this.props;
-    this.props.updateIsMobile(player.env.isTablet || player.env.isMobile || forceTouchUI);
+    const {isIPadOS, isTablet, isMobile} = player.env;
+    this.props.updateIsMobile(isIPadOS || isTablet || isMobile || forceTouchUI);
     this._onWindowResize();
     // TODO [es] check with Oren where they unlisten
     this.props.eventManager.listen(player, player.Event.RESIZE, () => this._onWindowResize());